+++ docs/src/components/ComponentDoc/ComponentExample/ComponentExample.tsx
@@ -64,7 +64,9 @@ class ComponentExample extends React.Component<ComponentExampleProps, ComponentE
   constructor(props) {
     super(props)
 
-    this.anchorName = examplePathToHash(props.examplePath)
+    const { examplePath } = props
+
+    this.anchorName = examplePathToHash(examplePath)
     this.state = {
       handleMouseLeave: this.handleMouseLeave,
       handleMouseMove: this.handleMouseMove,
@@ -72,7 +74,7 @@ class ComponentExample extends React.Component<ComponentExampleProps, ComponentE
       showCode: this.isActiveHash(),
       themeName: 'teams',
       componentVariables: {},
-      showRtl: false,
+      showRtl: examplePath && examplePath.endsWith('rtl') ? true : false,
       showTransparent: false,
       showVariables: false,
       isHovering: false,