+++ docs/src/examples/components/Button/index.tsx
@@ -4,6 +4,7 @@ import Variations from './Variations'
 import States from './States'
 import Groups from './Groups'
 import Usage from './Usage'
+import Rtl from './Rtl'
 import Performance from './Performance'
 
 const ButtonExamples = () => (
@@ -13,6 +14,7 @@ const ButtonExamples = () => (
     <Variations />
     <Groups />
     <Usage />
+    <Rtl />
     <Performance />
   </div>
 )